+++ components/src/core/drawer/drawer-body/nav-item/drawer-nav-item.component.ts
@@ -14,6 +14,7 @@ import {
 import { DrawerService } from '../../service/drawer.service';
 import { StateListener } from '../../state-listener.component';
 import { isEmptyView } from '../../../../utils/utils';
+import { MatExpansionPanel } from '@angular/material/expansion';
 
 export type DrawerNavItem = {
     statusColor?: string;
@@ -117,7 +118,7 @@ export type ActiveItemBackgroundShape = 'round' | 'square';
             </div>
             <!-- Nested Nav Items -->
             <mat-accordion displayMode="flat" class="blui-drawer-nested-nav-item" *ngIf="!isRail()">
-                <mat-expansion-panel [expanded]="expanded && isOpen()">
+                <mat-expansion-panel #matExpansionPanel>
                     <ng-content select="blui-drawer-nav-item"></ng-content>
                 </mat-expansion-panel>
             </mat-accordion>
@@ -180,6 +181,7 @@ export class DrawerNavItemComponent extends StateListener implements Omit<Drawer
     @ViewChild('icon') iconEl: ElementRef;
     @ViewChild('collapseIcon') collapseIconEl: ElementRef;
     @ViewChild('navItem') navItemEl: ElementRef;
+    @ViewChild('matExpansionPanel') matExpansionPanel: MatExpansionPanel;
 
     isEmpty = (el: ElementRef): boolean => isEmptyView(el);
 
@@ -193,6 +195,9 @@ export class DrawerNavItemComponent extends StateListener implements Omit<Drawer
     constructor(drawerService: DrawerService, changeDetectorRef: ChangeDetectorRef) {
         super(drawerService, changeDetectorRef);
         this.id = drawerService.createNavItemID();
+        this.drawerService.drawerOpenChanges().subscribe(() => {
+            this.handleExpand();
+        });
         this.drawerService.drawerActiveItemChanges().subscribe(() => {
             if (this.navItemEl) {
                 this.manageActiveItemTreeHighlight();
@@ -204,11 +209,30 @@ export class DrawerNavItemComponent extends StateListener implements Omit<Drawer
         if (changes.selected) {
             this.drawerService.emitChangeActiveItemEvent();
         }
+        if (changes.expanded !== undefined && !changes.expanded.isFirstChange() && this.matExpansionPanel) {
+            this.handleExpand();
+        }
+    }
+
+    ngAfterViewInit(): void {
+        this.handleExpand();
     }
 
     ngAfterContentInit(): void {
         // If ContentChildren is self-inclusive (ng version < 9), filter self out using service-generated NavItem ID.
         this.nestedNavItems = this.nestedNavItems.filter((item: DrawerNavItemComponent) => item.id !== this.id);
+    }
+
+    handleExpand(): void {
+        setTimeout(() => {
+            // Persistent drawers will only expand if they the drawer is opened.
+            // Temporary drawers will always have any expansion panels opened.
+            if (this.expanded && (this.isOpen() || this.drawerService.getDrawerVariant() === 'temporary')) {
+                this.matExpansionPanel.open();
+            } else {
+                this.matExpansionPanel.close();
+            }
+        });
     }
 
     isRail(): boolean {
@@ -296,5 +320,6 @@ export class DrawerNavItemComponent extends StateListener implements Omit<Drawer
 
     toggleNestedNavItems(): void {
         this.expanded = !this.expanded;
+        this.handleExpand();
     }
 }